Change to trait.
